Move ruler from TimeTrack to TimeTrackView
