+++ src/TimeTrack.cpp
@@ -19,9 +19,6 @@
 #include "ActiveProject.h"
 #include <cfloat>
 #include <wx/wxcrtvararg.h>
-#include <wx/dc.h>
-#include <wx/intl.h>
-#include "widgets/Ruler.h"
 #include "Envelope.h"
 #include "Mix.h"
 #include "Project.h"
@@ -72,11 +69,6 @@ void TimeTrack::CleanState()
 
    //Time track is always unique
    SetName(GetDefaultName());
-
-   mRuler = std::make_unique<Ruler>();
-   mRuler->SetUseZoomInfo(0, mZoomInfo);
-   mRuler->SetLabelEdges(false);
-   mRuler->SetFormat(Ruler::TimeFormat);
 }
 
 TimeTrack::TimeTrack(const TimeTrack &orig, ProtectedCreationArg &&a,
@@ -99,12 +91,6 @@ TimeTrack::TimeTrack(const TimeTrack &orig, ProtectedCreationArg &&a,
 
    mEnvelope->SetTrackLen( len );
    mEnvelope->SetOffset(0);
-
-   ///@TODO: Give Ruler:: a copy-constructor instead of this?
-   mRuler = std::make_unique<Ruler>();
-   mRuler->SetUseZoomInfo(0, mZoomInfo);
-   mRuler->SetLabelEdges(false);
-   mRuler->SetFormat(Ruler::TimeFormat);
 }
 
 // Copy the track metadata but not the contents.
